Updated to match app
Updated to address issues identified during integration with thehomebrewapp.

+++ thehomebrewapi/ValidationAttributes/BeerTypeIsValidEnumAttribute.cs
@@ -8,7 +8,7 @@ namespace thehomebrewapi.ValidationAttributes
     {
         public override bool IsValid(object value)
         {
-            var beerType = (int)value;
+            var beerType = (short)value;
 
             return Enum.IsDefined(typeof(ETypeOfBeer), beerType);
         }
